Move desolve to resolvers instead of utils
