+++ gundb/backends/utils.py
@@ -70,34 +70,3 @@ def defaultify(d):
 def eliminate_nones(lst):
     "Removes all Nonees in the given list"
     return [x for x in lst if x is not None]
-
-
-def desolve_obj(obj):
-    """Returns the given object in gundb form along with the souls it created"""
-    result = defaultify({})
-    added_souls = defaultify({})
-    for k, v in obj.items():
-        if k != METADATA and isinstance(v, dict):
-            prop_soul = v[METADATA][SOUL]
-            result[k] = {SOUL: prop_soul}
-            desolved_prop, added_in_prop = desolve_obj(v)
-            added_souls[prop_soul] = desolved_prop
-            for k, v in added_in_prop.items():
-                added_souls[k] = v
-        else:
-            result[k] = v
-    return result, added_souls
-
-
-def desolve(graph):
-    """resolve a graph in expanded form and convert it to gundb form"""   
-    result = defaultify({})
-    added_souls = defaultify({})
-    for k, v in graph.items():
-        prop_soul = v[METADATA][SOUL]
-        result[prop_soul], added_souls_in_obj = desolve_obj(v)
-        for k, v in added_souls_in_obj.items():
-            added_souls[k] = v
-    for k, v in added_souls.items():
-        result[k] = v
-    return result
